inlining MPIImplTraits and fallback into cpp file, minor mods addressing review comments
